# Using numba in line and multiline