--- a/openquake/hazardlib/geo/surface/multi.py
+++ b/openquake/hazardlib/geo/surface/multi.py
@@ -97,7 +97,7 @@ class MultiSurface(BaseSurface):
         """
         self.surfaces = surfaces
         self.tol = tol
-        self._set_tor()
+        self.tor = None
         self.areas = None
 
     # called at each instantiation
@@ -137,7 +137,7 @@ class MultiSurface(BaseSurface):
 
         # Set the multiline representing the rupture traces i.e. vertical
         # projections at the surface of the top of ruptures
-        self.tors = geo.MultiLine(tors)
+        self.tor = geo.MultiLine(tors)
 
     def get_min_distance(self, mesh):
         """
@@ -355,7 +355,9 @@ class MultiSurface(BaseSurface):
             A :class:`numpy.ndarray` instance with the Rx distance. Note that
             the Rx distance is directly taken from the GC2 t-coordinate.
         """
-        uut, tut = self.tors.get_tu(mesh)
+        if self.tor is None:
+            self._set_tor()
+        uut, tut = self.tor.get_tu(mesh)
         rx = tut[0] if len(tut[0].shape) > 1 else tut
         return rx
 
@@ -365,4 +367,6 @@ class MultiSurface(BaseSurface):
             An instance of :class:`openquake.hazardlib.geo.mesh.Mesh` with the
             coordinates of the sites.
         """
-        return self.tors.get_ry0_distance(mesh)
+        if self.tor is None:
+            self._set_tor()
+        return self.tor.get_ry0_distance(mesh)