This gives some speedup, but it is not expected to do much in many-site situations, since the problem of redundant calculations has not been tackled yet.
Still, the improvement in the test NZ2 in the oq-risk-tests is 6x in "total classical":
```
# branch get_tu
| calc_7471, maxmem=7.2 GB   | time_sec  | memory_mb | counts |
|----------------------------+-----------+-----------+--------|
| ClassicalCalculator.run    | 20.6      | 203.0     | 1      |
| total classical            | 16.1      | 9.08984   | 11     |
| total preclassical         | 6.87849   | 8.24219   | 16     |
| iter_ruptures              | 3.52899   | 1.96484   | 4_554  |

# master
| calc_7019, maxmem=7.9 GB   | time_sec  | memory_mb | counts |
|----------------------------+-----------+-----------+--------|
| total classical            | 97.6      | 39.2      | 11     |
| ClassicalCalculator.run    | 91.6      | 203.8     | 1      |
| iter_ruptures              | 64.3      | 120.0     | 4_554  |
| total preclassical         | 46.8      | 174.4     | 16     |
```
Notice also the spectacular improvement in `iter_ruptures` (18x). The situation for single site calculations looks much better, but there is still a lot to do. Here is the situation for the reduces USA model with 23,000+ sites:
```
# before
| calc_10, maxmem=73.8 GB    | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 61_182   | 233.0     | 159       |
| nonplanar contexts         | 45_211   | 0.0       | 10_448    |
| get_poes                   | 9_788    | 0.0       | 6_578_206 |
| computing mean_std         | 2_609    | 0.0       | 56_653    |
| ClassicalCalculator.run    | 2_170    | 847.8     | 1         |
| iter_ruptures              | 1_954    | 46.4      | 159       |
| total preclassical         | 363.0    | 12.5      | 53        |
| composing pnes             | 770.9    | 0.0       | 6_578_206 |

# after
| calc_12, maxmem=69.0 GB    | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 47_844   | 181.5     | 159       |
| nonplanar contexts         | 33_804   | 0.0       | 10_448    |
| get_poes                   | 9_740    | 0.0       | 6_578_206 |
| computing mean_std         | 2_563    | 0.0       | 56_653    |
| ClassicalCalculator.run    | 1_689    | 849.4     | 1         |
| composing pnes             | 759.2    | 0.0       | 6_578_206 |
| total preclassical         | 301.0    | 11.3      | 53        |
| iter_ruptures              | 141.7    | 1.15104   | 159       |
```